♻️  router utils.py 추출 — env / coercion / message parsing (P0-P step 4)
변경 이유
- env reader / 값 coercion / discord-agnostic message parsing / async
  helper / F16 recall coverage 가 한 줄짜리 유틸인데 ``_legacy`` 안에
  흩어져 있었다. 모든 후속 모듈이 의존하므로 leaf 로 분리.

주요 변경 사항
- 새 파일 ``engineering_channel_router/utils.py`` (188 줄):
  - 값 coercion: ``_optional_str``, ``_safe_int``,
    ``_normalize_channel_name``
  - env reader: ``_optional_int_env``, ``_optional_string_env``,
    ``_optional_bool_env``
  - discord-agnostic message parsing: ``extract_user_links_from_message``,
    ``extract_message_attachments``
  - async helper: ``_maybe_await``
  - F16 recall coverage enrichment: ``_attach_recall_coverage``
- ``_legacy.py`` 의 utility 2 블록 (16 + 108 = 124 줄) 제거 → re-export.
- ``models.EngineeringRouteContext.configured`` / ``.from_env`` 의 lazy
  import 을 ``._legacy`` → ``.utils`` 로 정리. step 3 의 임시 우회 제거.
- ``__init__.py`` facade 가 ``.utils`` 에서 직접 import.

비고
- 4672 PASS + 1 skipped 회귀 유지.
