+++ src/yule_orchestrator/discord/engineering_channel_router/_legacy.py
@@ -2997,20 +2997,8 @@ def _coerce_research_loop_report(raw: Any) -> EngineeringResearchLoopReport:
     )
 
 
-def _optional_str(value: Any) -> Optional[str]:
-    if value is None:
-        return None
-    text = str(value).strip()
-    return text or None
-
-
-def _safe_int(value: Any) -> Optional[int]:
-    if value is None:
-        return None
-    try:
-        return int(value)
-    except (TypeError, ValueError):
-        return None
+# P0-P step 4: value coercion helpers extracted to .utils.
+from .utils import _optional_str, _safe_int  # noqa: E402,F401 — re-export
 
 
 def _coerce_outcome(
@@ -3062,111 +3050,14 @@ def _coerce_outcome(
     )
 
 
-async def _maybe_await(value: Any) -> Any:
-    if hasattr(value, "__await__"):
-        return await value
-    return value
-
-
-def extract_user_links_from_message(
-    message: Any,
-    prompt_text: Optional[str] = None,
-) -> tuple[str, ...]:
-    """Pull URLs out of the user's message body.
-
-    Lazily delegates to :func:`research_collector.extract_urls` so we get
-    the same regex + dedup the collector uses internally. Returns an empty
-    tuple if the helper isn't importable (e.g. during a partial install).
-    """
-
-    text = (prompt_text or getattr(message, "content", "") or "")
-    if not text:
-        return ()
-    try:
-        from ...agents.research.collector import extract_urls
-    except Exception:  # noqa: BLE001
-        return ()
-    return tuple(extract_urls(text))
-
-
-def extract_message_attachments(message: Any) -> tuple[Any, ...]:
-    """Return the message's attachments as a stable tuple, discord.py-agnostic.
-
-    discord.py exposes ``message.attachments`` as a list of ``Attachment``
-    objects, but tests pass plain dataclasses or dicts. We accept any iterable
-    and drop ``None`` entries so the engineering conversation layer can rely
-    on a clean sequence regardless of the Discord shape.
-    """
-
-    raw = getattr(message, "attachments", None)
-    if raw is None:
-        return ()
-    if isinstance(raw, (list, tuple)):
-        return tuple(item for item in raw if item is not None)
-    try:
-        return tuple(item for item in raw if item is not None)
-    except TypeError:
-        return ()
-
-
-def _normalize_channel_name(value: object | None) -> str:
-    if value is None:
-        return ""
-    return str(value).strip().lstrip("#").lower()
-
-
-def _optional_int_env(name: str) -> Optional[int]:
-    raw = os.environ.get(name)
-    if raw is None or not raw.strip():
-        return None
-    value = raw.strip()
-    try:
-        return int(value)
-    except ValueError as exc:
-        raise ValueError(
-            f"{name} must be an integer value, got: {value!r}"
-        ) from exc
-
-
-def _optional_string_env(name: str) -> Optional[str]:
-    raw = os.environ.get(name)
-    if raw is None:
-        return None
-    value = raw.strip()
-    return value or None
-
-
-def _optional_bool_env(name: str, *, default: bool = False) -> bool:
-    """Parse a boolean envvar — empty/unset returns ``default``.
-
-    Accepted truthy values: ``"true"``, ``"1"``, ``"yes"``, ``"on"``
-    (case-insensitive). Anything else is treated as the default. Used
-    by F16 ``EngineeringRouteContext.prefer_recall_first_gateway`` so
-    operators can opt into the recall-first gateway path without code
-    changes.
-    """
-
-    raw = os.environ.get(name)
-    if raw is None:
-        return default
-    value = raw.strip().lower()
-    if not value:
-        return default
-    return value in {"true", "1", "yes", "on"}
-
-
-def _attach_recall_coverage(recall: RuntimeRecallResult) -> RuntimeRecallResult:
-    """F16 — replace ``recall`` with a copy whose ``coverage`` is scored.
-
-    The scorer is **defensive**: any failure (None, malformed hits)
-    degrades to ``RecallCoverage(level=low, stale=True)``. Legacy
-    callers that ignore ``coverage`` see no behaviour change.
-    """
-
-    try:
-        coverage = compute_recall_coverage(recall)
-    except Exception:  # noqa: BLE001
-        coverage = RecallCoverage(
-            level="low", stale=True, sources=(), reason="scorer raised"
-        )
-    return replace(recall, coverage=coverage)
+# P0-P step 4: async + message-parsing + env + recall coverage extracted to .utils.
+from .utils import (  # noqa: E402,F401 — re-export for back-compat
+    _attach_recall_coverage,
+    _maybe_await,
+    _normalize_channel_name,
+    _optional_bool_env,
+    _optional_int_env,
+    _optional_string_env,
+    extract_message_attachments,
+    extract_user_links_from_message,
+)
